requirements.txt
rich>=14.1.0
pytest>=8.4.2
pytest-cov>=7.0.0
pyright>=1.1.405
pyyaml>=6.0.2
pexpect>=4.8.0
prompt_toolkit>=3.0.36
